requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==1.0.0
Django==4.0.6
django-bootstrap-v5==1.0.11
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
newsapi-python==0.2.6
psycopg2==2.9.3
python-decouple==3.6
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
